+++ app/routes/reservation.create.tsx
@@ -1,37 +1,15 @@
 import { Form, useLoaderData } from "@remix-run/react";
 import { FormEventHandler, useEffect, useState } from "react";
-import { getById, post } from "~/api/reservation";
+import { post } from "~/api/reservation";
 import { Reservation } from "~/components/Reservation";
 import { toDatetimeLocal } from "~/utils/datetime";
 
-export const loader = async ({ params }:any) => {
-    return getById(params.id).then((res) => {
-        if (res == undefined) {
-            console.error("No reservation found");
-            return {"reservation": undefined, "getError": "No reservation found"};
-        }
-        return {"reservation": res, "getError": undefined};
-    });
-  
-};
-
 export default function EditReservation() {
     //displays a react component that allows the user to edit a reservation
-    const {reservation, getError} = useLoaderData<typeof loader>();
-    
     const [title, setTitle] = useState("");
     const [room, setRoom] = useState("");
     const [start, setStart] = useState<Date>(new Date());
     const [end, setEnd] = useState<Date>(new Date());
-
-    useEffect(() => {
-        if (reservation != undefined) {
-            setTitle(reservation.title);
-            setRoom(reservation.room);
-            setStart(reservation.start);
-            setEnd(reservation.end);
-        }
-    }, [reservation]);
 
     const handleChange: FormEventHandler<HTMLFormElement> = (event: any) => {
         switch(event.target.title) {
@@ -58,11 +36,7 @@ export default function EditReservation() {
     const handleSubmit: FormEventHandler<HTMLFormElement> = (event: any) => {
         event.preventDefault();
         console.log(title, room, start, end);
-        if (!reservation) {
-            console.error("No reservation found");
-            return;
-        }
-        post(new Reservation(reservation.id, title, room, start, end)).then((res) => {
+        post(new Reservation(-1, title, room, start, end)).then((res) => {
             
         });
 
@@ -70,8 +44,8 @@ export default function EditReservation() {
 
     return (
         <div>
-            <h1 key="title">Edit Reservation</h1>
-            <Form method="PUT" onChange={handleChange} onSubmit={handleSubmit}>
+            <h1 key="title">Create Reservation</h1>
+            <Form method="POST" onChange={handleChange} onSubmit={handleSubmit}>
                 <input title="title" name="title" type="text" defaultValue={title}/>
                 <input title="room" name="room" type="text" defaultValue={room}/>
                 <input title="start" name="start" type="datetime-local" defaultValue={toDatetimeLocal(start)}/>
